+++ CMakeLists.txt
@@ -15,15 +15,16 @@ set(Boost_USE_MULTITHREADED ON)
 find_package(Boost REQUIRED)
 include_directories(${Boost_INCLUDE_DIRS})
 find_package(Threads REQUIRED)
-set(SOURCE src/router.cpp src/handler_classes.cpp src/query_creater.cpp src/req_parser.cpp nlohmann/json.hpp src/common_chat.cpp src/message_chat_view.cpp src/user_chat_view.cpp)
+set(SOURCE nlohmann/json.hpp
+        src/CommonChatView.cpp src/MessageChatView.cpp src/UserChatView.cpp include/ResponseCreater.h include/View.h)
 include_directories(include)
 include_directories(nlohmann)
-add_executable(server ${SOURCE})
-target_sources(server PUBLIC src/chat_server.cpp)
+#add_executable(server ${SOURCE})
+#target_sources(server PUBLIC src/chat_server.cpp)
 link_libraries(mysqlcppconn Threads::Threads ${Boost_LIBRARIES})
 ############test generate
 find_package(GTest REQUIRED)
-add_executable(Test tests/test.cpp)
+add_executable(Test tests/tests.cpp)
 target_include_directories(Test PUBLIC ${GTEST_INCLUDE_DIRS})
 target_link_libraries(Test PUBLIC ${GTEST_LIBRARIES} Threads::Threads)
 include_directories(/usr/include/cppconn)
