--- include/query_creater.h
@@ -1,11 +1,0 @@
-#ifndef PLUS_PROJECT_QUERY_CREATER_H
-#define PLUS_PROJECT_QUERY_CREATER_H
-
-#include "connector.h"
-#include <cstring>
-
-class Query_creater {
-};
-
-
-#endif //PLUS_PROJECT_QUERY_CREATER_H
