+++ src/CommonChatView.cpp
@@ -1,0 +1,2 @@
+#include "CommonChatView.h"
+
