+++ src/MessageChatView.cpp
@@ -1,0 +1,3 @@
+
+#include "MessageChatView.h"
+
