+++ src/UserChatView.cpp
@@ -1,0 +1,2 @@
+
+#include "UserChatView.h"
