--- src/chat.cpp
@@ -1,3 +1,0 @@
-#include "../include/chat.h"
-
-shared_ptr<vector<int>> OnlineChat::getpeople() {}
