--- src/common_chat.cpp
@@ -1,2 +1,0 @@
-#include "common_chat.h"
-
