--- src/creator_fabr_method.cpp
@@ -1,1 +1,0 @@
-#include "../include/creator_fabr_method.h"
