--- src/message_chat_view.cpp
@@ -1,3 +1,0 @@
-
-#include "message_chat_view.h"
-
