--- src/query_creater.cpp
@@ -1,1 +1,0 @@
-#include "query_creater.h"
