--- src/user_chat_view.cpp
@@ -1,3 +1,0 @@
-
-
-#include "user_chat_view.h"
